Block Functionality: * Added Infura integration for Block-related requests * Added Infura models and parsing logic * Fixed querystring logic for Infura requests * Added business logic to BlockService to fulfill endpoint needs * Added block controller with all necessary endpoints
Health Checks:
* Added bare-bones health check endpoint

Infura Proxy:
* Added infura proxy logic for all requests unhandled by this application

+++ src/main/java/com/meisterassociates/apicache/model/GasPrice.java
@@ -9,15 +9,15 @@ import java.util.Map;
 
 public class GasPrice extends CacheableModel {
 
-    private final int id;
+    private final long id;
     private final String jsonrpc;
     private final BigInteger priceInWei;
 
-    public GasPrice(int id, String jsonrpc, BigInteger priceInWei) {
+    public GasPrice(long id, String jsonrpc, BigInteger priceInWei) {
         this(id, jsonrpc, priceInWei, LocalDateTime.now());
     }
 
-    public GasPrice(int id, String jsonrpc, BigInteger priceInWei, LocalDateTime datetime) {
+    public GasPrice(long id, String jsonrpc, BigInteger priceInWei, LocalDateTime datetime) {
         super(datetime);
 
         this.id = id;
